third_party/WebKit/LayoutTests/fast/events/pointerevents/touch-pointercancel.html

<!DOCTYPE HTML>
<script src="../../../resources/js-test.js"></script>
<style>
  div.test {
    margin: 10px;
    width: 100px;
    height: 200px;
    overflow: scroll;
    border: solid 1px;
    float: left;
  }
</style>

<div id="target" class="test"></div>

<div id="console"></div>

<script>
description("Verifies that pointercancel events are fired correctly on touch.");

var eventToPreventDefault = "";

function init() {
  var eventList = ["pointerdown", "pointerup", "pointermove", "pointercancel", "pointerover", "pointerenter", "pointerout", "pointerleave",
                   "touchstart", "touchend", "touchmove", "touchcancel"];

  eventList.forEach(function(eventName) {
    document.getElementById("target").addEventListener(eventName, function(event) {
      var idMsg;
      if (event.type.startsWith("pointer"))
        idMsg = ", pointerId=" + event.pointerId;
      else {
        idMsg = ", touchIds=";
        touches = event.changedTouches;
        for (var i = 0; i < touches.length; ++i) {
          idMsg += touches[i].identifier;
        };
      }

      var preventDefaultMsg = "";

      if (event.type == eventToPreventDefault) {
        event.preventDefault();
        preventDefaultMsg = ", prevented default";
      }

      debug("Received " + event.type + idMsg + preventDefaultMsg);
    });
  });
}

function testSingleDrag(notifyStartOfTouchScroll) {
  var rect = document.getElementById("target").getBoundingClientRect();
  var x = rect.left + 50;
  var y = rect.top + 50;

  eventSender.addTouchPoint(x, y);
  eventSender.touchStart();

  eventSender.updateTouchPoint(0, x, y+50);
  eventSender.touchMove();

  if (notifyStartOfTouchScroll)
    eventSender.notifyStartOfTouchScroll();

  eventSender.updateTouchPoint(0, x, y+100);
  eventSender.touchMove();

  eventSender.updateTouchPoint(0, x, y+150);
  eventSender.touchMove();

  eventSender.releaseTouchPoint(0);
  eventSender.touchEnd();
}

function testMultiDrag(notifyStartOfTouchScroll) {
  var rect;

  rect = document.getElementById("target").getBoundingClientRect();
  var x0 = rect.left + 30;
  var x1 = rect.left + 50;
  var x2 = rect.left + 70;
  var x3 = rect.left + 55;
  var y = rect.top + 50;

  eventSender.addTouchPoint(x0, y);
  eventSender.addTouchPoint(x1, y);
  eventSender.addTouchPoint(x2, y);
  eventSender.touchStart();

  y += 50;
  eventSender.updateTouchPoint(0, x0, y);
  eventSender.updateTouchPoint(1, x1, y);
  eventSender.updateTouchPoint(2, x2, y);
  eventSender.touchMove();

  y += 50;
  eventSender.updateTouchPoint(0, x0, y);
  eventSender.updateTouchPoint(1, x1, y);
  eventSender.updateTouchPoint(2, x2, y);
  eventSender.touchMove();

  if (notifyStartOfTouchScroll)
    eventSender.notifyStartOfTouchScroll();

  eventSender.addTouchPoint(x3, y);
  eventSender.touchStart();

  y += 50;
  eventSender.updateTouchPoint(0, x0, y);
  eventSender.updateTouchPoint(1, x1, y);
  eventSender.updateTouchPoint(2, x2, y);
  eventSender.updateTouchPoint(3, x3, y);
  eventSender.touchMove();

  eventSender.releaseTouchPoint(0);
  eventSender.releaseTouchPoint(1);
  eventSender.releaseTouchPoint(2);
  eventSender.releaseTouchPoint(3);
  eventSender.touchEnd();
}

function testMultiCancel() {
  var rect;

  rect = document.getElementById("target").getBoundingClientRect();
  var x0 = rect.left + 30;
  var x1 = rect.left + 50;
  var x2 = rect.left + 70;
  var y = rect.top + 50;

  eventSender.addTouchPoint(x0, y);
  eventSender.addTouchPoint(x1, y);
  eventSender.addTouchPoint(x2, y);
  eventSender.touchStart();

  eventSender.cancelTouchPoint(2);
  eventSender.touchCancel();

  y += 50;
  eventSender.updateTouchPoint(0, x0, y);
  eventSender.updateTouchPoint(1, x2, y);
  eventSender.touchMove();

  eventSender.cancelTouchPoint(0);
  eventSender.cancelTouchPoint(1);
  eventSender.touchCancel();
}

function runTests() {
  debug("--- single drag tests ---");
  debug("");

  debug("-- preventDefault on none, without notifyStartOfTouchScroll --");
  eventToPreventDefault = "-none-";
  testSingleDrag(false);
  debug("");

  debug("-- preventDefault on pointermove, without notifyStartOfTouchScroll --");
  eventToPreventDefault = "pointermove";
  testSingleDrag(false);
  debug("");

  debug("-- preventDefault on none, with notifyStartOfTouchScroll --");
  eventToPreventDefault = "-none-";
  testSingleDrag(true);
  debug("");

  debug("-- preventDefault on pointermove, with notifyStartOfTouchScroll --");
  eventToPreventDefault = "pointermove";
  testSingleDrag(true);
  debug("");

  debug("--- multi drag tests ---");
  debug("");

  debug("-- preventDefault on none, without notifyStartOfTouchScroll --");
  eventToPreventDefault = "-none-";
  testMultiDrag(false);
  debug("");

  debug("-- preventDefault on none, with notifyStartOfTouchScroll --");
  eventToPreventDefault = "-none-";
  testMultiDrag(true);
  debug("");

  debug("--- multi cancel tests ---");
  debug("");
  testMultiCancel();
  debug("");
}


init();
if (window.eventSender)
  runTests();
else
  debug("This test requires eventSender");

</script>
