third_party/WebKit/LayoutTests/fast/events/pointerevents/touch-pointercancel-expected.txt

Verifies that pointercancel events are fired correctly on touch.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


--- single drag tests ---

-- preventDefault on none, without notifyStartOfTouchScroll --
Received pointerover, pointerId=2
Received pointerenter, pointerId=2
Received pointerdown, pointerId=2
Received touchstart, touchIds=0
Received pointermove, pointerId=2
Received touchmove, touchIds=0
Received pointermove, pointerId=2
Received touchmove, touchIds=0
Received pointermove, pointerId=2
Received touchmove, touchIds=0
Received pointerup, pointerId=2
Received pointerout, pointerId=2
Received pointerleave, pointerId=2
Received touchend, touchIds=0

-- preventDefault on pointermove, without notifyStartOfTouchScroll --
Received pointerover, pointerId=3
Received pointerenter, pointerId=3
Received pointerdown, pointerId=3
Received touchstart, touchIds=0
Received pointermove, pointerId=3, prevented default
Received touchmove, touchIds=0
Received pointermove, pointerId=3, prevented default
Received touchmove, touchIds=0
Received pointermove, pointerId=3, prevented default
Received touchmove, touchIds=0
Received pointerup, pointerId=3
Received pointerout, pointerId=3
Received pointerleave, pointerId=3
Received touchend, touchIds=0

-- preventDefault on none, with notifyStartOfTouchScroll --
Received pointerover, pointerId=4
Received pointerenter, pointerId=4
Received pointerdown, pointerId=4
Received touchstart, touchIds=0
Received pointermove, pointerId=4
Received touchmove, touchIds=0
Received pointercancel, pointerId=4
Received pointerout, pointerId=4
Received pointerleave, pointerId=4
Received touchmove, touchIds=0
Received touchmove, touchIds=0
Received touchend, touchIds=0

-- preventDefault on pointermove, with notifyStartOfTouchScroll --
Received pointerover, pointerId=5
Received pointerenter, pointerId=5
Received pointerdown, pointerId=5
Received touchstart, touchIds=0
Received pointermove, pointerId=5, prevented default
Received touchmove, touchIds=0
Received pointercancel, pointerId=5
Received pointerout, pointerId=5
Received pointerleave, pointerId=5
Received touchmove, touchIds=0
Received touchmove, touchIds=0
Received touchend, touchIds=0

--- multi drag tests ---

-- preventDefault on none, without notifyStartOfTouchScroll --
Received pointerover, pointerId=6
Received pointerenter, pointerId=6
Received pointerdown, pointerId=6
Received pointerover, pointerId=7
Received pointerenter, pointerId=7
Received pointerdown, pointerId=7
Received pointerover, pointerId=8
Received pointerenter, pointerId=8
Received pointerdown, pointerId=8
Received touchstart, touchIds=012
Received pointermove, pointerId=6
Received pointermove, pointerId=7
Received pointermove, pointerId=8
Received touchmove, touchIds=012
Received pointermove, pointerId=6
Received pointermove, pointerId=7
Received pointermove, pointerId=8
Received touchmove, touchIds=012
Received pointerover, pointerId=9
Received pointerenter, pointerId=9
Received pointerdown, pointerId=9
Received touchstart, touchIds=3
Received pointermove, pointerId=6
Received pointermove, pointerId=7
Received pointermove, pointerId=8
Received pointermove, pointerId=9
Received touchmove, touchIds=0123
Received pointerup, pointerId=6
Received pointerout, pointerId=6
Received pointerleave, pointerId=6
Received pointerup, pointerId=7
Received pointerout, pointerId=7
Received pointerleave, pointerId=7
Received pointerup, pointerId=8
Received pointerout, pointerId=8
Received pointerleave, pointerId=8
Received pointerup, pointerId=9
Received pointerout, pointerId=9
Received pointerleave, pointerId=9
Received touchend, touchIds=0123

-- preventDefault on none, with notifyStartOfTouchScroll --
Received pointerover, pointerId=10
Received pointerenter, pointerId=10
Received pointerdown, pointerId=10
Received pointerover, pointerId=11
Received pointerenter, pointerId=11
Received pointerdown, pointerId=11
Received pointerover, pointerId=12
Received pointerenter, pointerId=12
Received pointerdown, pointerId=12
Received touchstart, touchIds=012
Received pointermove, pointerId=10
Received pointermove, pointerId=11
Received pointermove, pointerId=12
Received touchmove, touchIds=012
Received pointermove, pointerId=10
Received pointermove, pointerId=11
Received pointermove, pointerId=12
Received touchmove, touchIds=012
Received pointercancel, pointerId=10
Received pointerout, pointerId=10
Received pointerleave, pointerId=10
Received pointercancel, pointerId=11
Received pointerout, pointerId=11
Received pointerleave, pointerId=11
Received pointercancel, pointerId=12
Received pointerout, pointerId=12
Received pointerleave, pointerId=12
Received touchstart, touchIds=3
Received touchmove, touchIds=0123
Received touchend, touchIds=0123

--- multi cancel tests ---

Received pointerover, pointerId=13
Received pointerenter, pointerId=13
Received pointerdown, pointerId=13
Received pointerover, pointerId=14
Received pointerenter, pointerId=14
Received pointerdown, pointerId=14
Received pointerover, pointerId=15
Received pointerenter, pointerId=15
Received pointerdown, pointerId=15
Received touchstart, touchIds=012
Received pointercancel, pointerId=15
Received pointerout, pointerId=15
Received pointerleave, pointerId=15
Received touchcancel, touchIds=2
Received pointermove, pointerId=13
Received pointermove, pointerId=14
Received touchmove, touchIds=01
Received pointercancel, pointerId=13
Received pointerout, pointerId=13
Received pointerleave, pointerId=13
Received pointercancel, pointerId=14
Received pointerout, pointerId=14
Received pointerleave, pointerId=14
Received touchcancel, touchIds=01

PASS successfullyParsed is true

TEST COMPLETE
